+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Yii2 multiple input change log
 ==============================
 
+2.6.0
+=====
+
+- PR#132: Implemented `Sortting` (sankam-nikolya)
+- PR#132: fixed if attribute is set and hasProperty return false (sankam-nikolya)
+
 2.5.0
 =====